+++ src/Telegram.Bot/Types/Enums/MessageType.cs
@@ -77,6 +77,16 @@ namespace Telegram.Bot.Types.Enums
         /// The <see cref="Message"/> contains a <see cref="VideoNote"/>
         /// </summary>
         VideoNoteMessage,
+
+        /// <summary>
+        /// The <see cref="Message"/> contains a <see cref="Invoice"/>
+        /// </summary>
+        Invoice,
+
+        /// <summary>
+        /// The <see cref="Message"/> contains a <see cref="SuccessfulPayment"/>
+        /// </summary>
+        SuccessfulPayment,
     }
 
     internal static class MessageTypeExtension
@@ -109,7 +119,8 @@ namespace Telegram.Bot.Types.Enums
                     return new KeyValuePair<string, string>("sendGame", "game_short_name");
                 case MessageType.VideoNoteMessage:
                     return new KeyValuePair<string, string>("sendVideoNote", "video_note");
-
+                case MessageType.Invoice:
+                    return new KeyValuePair<string, string>("sendInvoice", "title");
                 default:
                     throw new NotImplementedException();
             }