+++ src/Telegram.Bot/Types/Payments/ShippingOption.cs
@@ -1,0 +1,29 @@
+﻿using Newtonsoft.Json;
+
+namespace Telegram.Bot.Types.Payments
+{
+    /// <summary>
+    /// This object represents one shipping option.
+    /// </summary>
+    [JsonObject(MemberSerialization.OptIn)]
+    public class ShippingOption
+    {
+        /// <summary>
+        /// Shipping option identifier
+        /// </summary>
+        [JsonProperty("id")]
+        public string Id { get; set; }
+
+        /// <summary>
+        /// Option title
+        /// </summary>
+        [JsonProperty("title")]
+        public string Title { get; set; }
+
+        /// <summary>
+        /// List of price portions
+        /// </summary>
+        [JsonProperty("prices")]
+        public LabeledPrice[] Prices { get; set; } 
+    }
+}